% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom_classifier.R
\name{Custom_classifier}
\alias{Custom_classifier}
\title{Add custom training datasets to perform prediction}
\usage{
Custom_classifier(testdata, cores, CustomDataList)
}
\arguments{
\item{testdata}{Cell type to be predicted.
(Note: Gene expression values of the samples need to be scaled to the range [0,1] by the function \link{PreCanCell_data} first)}

\item{cores}{Number of cores for parallel computing.}

\item{CustomDataList}{A list of training datasets(added to TrainDataList).}
}
\value{
A dataframe with 4 columns:
\item{Sample}{Cell id}
\item{freq_cancer}{the ratio of the number of classifiers predicting that the cell is a cancer cell to 5 classifiers}
\item{freq_non_cancer}{the ratio of the number of classifiers predicting that the cell is a non-cancer cell to 5 classifiers}
\item{pred_labels}{Prediction results}
}
\description{
\code{Custom_classifier} Preprocessing for custom training datasets.
}
