% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom_preprocessing.R
\name{Custom_preprocessing}
\alias{Custom_preprocessing}
\title{Add custom training datasets to perform prediction}
\usage{
Custom_preprocessing(
  counts,
  annotation,
  project = "PanCanCell",
  min.cells = 400,
  min.features = 0,
  normalization.method = "LogNormalize",
  scale.factor = 10000,
  verbose = TRUE
)
}
\arguments{
\item{counts}{The counts matrix of scRNA, the rowname is GeneSymbol, the colname is cell names.}

\item{annotation}{The annotation of scRNA with two columns, first column is cell name(matched with counts),
second column is cell type (tumor or non-tumor).}
}
\value{
A dataframe
}
\description{
\code{Custom_preprocessing} Preprocessing for custom training datasets.
}
