% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\docType{data}
\name{feature}
\alias{feature}
\title{Feature}
\format{
A data frame with 259 rows and 2 columns. The column names
are as follows:
\describe{
  \item{Symbol}{Official Gene Symbol}
  \item{ID}{Entrez Gene ID}
}
}
\source{
The 259 gene features were selected through five single-cell transcriptome datasets.
}
\usage{
feature
}
\description{
The features have significant associations with cancer-related molecular and clinical features in pan-cancer.
}
\keyword{datasets}
