% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kNN1.R
\docType{data}
\name{kNN1}
\alias{kNN1}
\title{Training sets}
\format{
A k-NN (k = 5) model.
}
\source{
\url{https://doi.org/10.1016/j.ccell.2021.02.015}
}
\usage{
kNN1
}
\description{
Training sets
}
\keyword{model}
