% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kNN2.R
\docType{data}
\name{kNN2}
\alias{kNN2}
\title{Training sets}
\format{
A k-NN (k = 5) model.
}
\source{
\url{https://doi.org/10.1016/j.cell.2017.10.044}
}
\usage{
kNN2
}
\description{
Training sets
}
\keyword{model}
