% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kNN3.R
\docType{data}
\name{kNN3}
\alias{kNN3}
\title{Training sets}
\format{
A k-NN (k = 5) model.
}
\source{
\url{https://doi.org/10.1016/j.cell.2018.09.006}
}
\usage{
kNN3
}
\description{
Training sets
}
\keyword{model}
