% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kNN4.R
\docType{data}
\name{kNN4}
\alias{kNN4}
\title{Training sets}
\format{
A k-NN (k = 5) model.
}
\source{
\url{https://doi.org/10.1038/s41467-020-16164-1}
}
\usage{
kNN4
}
\description{
Training sets
}
\keyword{model}
