% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kNN5.R
\docType{data}
\name{kNN5}
\alias{kNN5}
\title{Training sets}
\format{
A k-NN (k = 5) model.
}
\source{
\url{https://doi.org/10.1038/s41588-021-00911-1}
}
\usage{
kNN5
}
\description{
Training sets
}
\keyword{model}
