% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreCanCell_classifier.R
\name{PreCanCell_classifier}
\alias{PreCanCell_classifier}
\title{Predicting Cancer and Non-Cancer Cells from Single-cell Transcriptomes}
\usage{
PreCanCell_classifier(testdata, cores)
}
\arguments{
\item{testdata}{Cell type to be predicted.
(Note: Gene expression values of the samples need to be scaled to the range [0,1] by the function \link{PreCanCell_data} first)}

\item{cores}{Number of cores for parallel computing.}
}
\value{
A dataframe with 2 columns:
\item{Sample}{Cell id}
\item{pred_labels}{Prediction results}
}
\description{
\code{PreCanCell_classifier} Identify malignant and non-malignant cells across cancer types.
}
\examples{
path <- system.file("extdata", "example.txt", package = "PreCanCell", mustWork = TRUE)
input <- read.table(path, stringsAsFactors = FALSE, header = TRUE, check.names = FALSE, sep = "\t", quote = "", row.names = 1)
testdata <- PreCanCell_data(input)
results <- PreCanCell_classifier(testdata, 2)
}
