% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreCanCell__classifier2.R
\name{PreCanCell_classifier2}
\alias{PreCanCell_classifier2}
\title{Predicting Cancer and Non-Cancer Cells from Single-cell Transcriptomes}
\usage{
PreCanCell_classifier2(testdata, cores)
}
\arguments{
\item{testdata}{Cell type to be predicted.
(Note: Gene expression values of the samples need to be scaled to the range [0,1] by the function \link{PreCanCell_data} first)}

\item{cores}{Number of cores for parallel computing.}
}
\value{
A dataframe with 2 columns:
\item{Sample}{Cell id}
\item{pred_labels}{Prediction results}
}
\description{
\code{PreCanCell_classifier2} Identify malignant and non-malignant cells across cancer types.
}
\examples{
path <- system.file("extdata", "example.txt", package = "PreCanCell", mustWork = TRUE)
input <- load(path)
testdata <- PreCanCell_data(input)
results <- PreCanCell_classifier2(testdata, 2)
}
