% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreCanCell_data.R
\name{PreCanCell_data}
\alias{PreCanCell_data}
\title{Data Preprocessing}
\usage{
PreCanCell_data(input)
}
\arguments{
\item{input}{The normalized expression dataframe, with rows being genes, and columns being cells.}
}
\value{
A m*n matrix contains the expression values of n matched genes in m cells.
}
\description{
\code{PreCanCell_data} This function selects matched genes and scales gene expression values (maximum=1, minimum=0).
}
\details{
The "input" refers to a normalized expression data with gene symbol.
Please note that \code{feature} must be present in row of \code{input} and
should not contain any duplicated feature names.
}
\examples{
path <- system.file("extdata", "example.txt", package = "PreCanCell", mustWork = TRUE)
input <- read.table(path, stringsAsFactors = FALSE, header = TRUE, check.names = FALSE, sep = "\t", quote = "", row.names = 1)
testdata <- PreCanCell_data(input)
}
