% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreCanCellPre.R
\docType{package}
\name{PreCanCellPre-package}
\alias{PreCanCellPre-package}
\alias{PreCanCellPre}
\title{Prediction of cancer and non-cancer cells from single-cell transcriptomes}
\description{
A single cell was classified as malignant or non-malignant based on its gene expression profile by the k-NN (k = 5) classifier.
}
