% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainDataList.R
\docType{data}
\name{TrainDataList}
\alias{TrainDataList}
\title{List of training sets}
\format{
A list with 5 training sets.
}
\source{
\url{https://doi.org/10.1016/j.ccell.2021.02.015}
\url{https://doi.org/10.1016/j.cell.2017.10.044}
\url{https://doi.org/10.1016/j.cell.2018.09.006}
\url{https://doi.org/10.1038/s41467-020-16164-1}
\url{https://doi.org/10.1038/s41588-021-00911-1}
}
\usage{
TrainDataList
}
\description{
The list of training sets are the expression profiles of a 295-gene panel from scRNA-Seq pan-cancer
(involving RCC, HNSCC, Melanoma, LUAD, and BC cancers) datasets.
}
\keyword{datasets}
